requirements.txt
# Python == 3.8.10
# uvicorn == 0.17.6
fastapi==0.78.0
python-jose==3.3.0
SQLAlchemy==1.4.26
passlib==1.7.4